To long to handle in only 1 comment
I had to fixed a lot of problems guys. We have already talk about that

+++ vote-sys/src/main/java/com/senatic/votesys/model/dto/CandidatoDTO.java
@@ -11,6 +11,6 @@ public class CandidatoDTO {
     
     private String documento;
     private MultipartFile imagen;
-    private Long idVotacion;
+    private Integer idVotacion;
     private String propuestas;
 }
